+++ server/src/routes/index.ts
@@ -1,9 +1,9 @@
-var express = require('express');
+import express from 'express';
 var router = express.Router();
 
 /* GET home page. */
-router.get('/', function(req, res, next) {
+router.get('/', function(_req, res, _next) {
   res.render('index', { title: 'Express' });
 });
 
-module.exports = router;
+export default router;
